Added fish config
Most of this is autogenerated.

he most important thing here is actually the one that's probably going
to change: the ssh.fish file
This was useful when I used kitty, but now that my main terminal
emulator is Ghostty, this has kinda become useless.

'm hoping Ghostty's devs implement something like kitten, which would be
really dope. In this case, I'll just update this file to their version
of it.